fix: update 17 demo examples to use typed Market constants (P6-9)

+++ examples/12_capital_flow/main.go
@@ -22,7 +22,7 @@ func main() {
 		log.Fatalf("Connect failed: %v", err)
 	}
 
-	flows, err := client.GetCapitalFlow(context.Background(), cli, int32(constant.Market_HK), "00700")
+	flows, err := client.GetCapitalFlow(context.Background(), cli, constant.Market_HK, "00700")
 	if err != nil {
 		log.Fatalf("GetCapitalFlow failed: %v", err)
 	}
