+++ examples/13_plate_set/main.go
@@ -22,7 +22,7 @@ func main() {
 		log.Fatalf("Connect failed: %v", err)
 	}
 
-	plates, err := client.GetPlateSet(context.Background(), cli, int32(constant.Market_US))
+	plates, err := client.GetPlateSet(context.Background(), cli, constant.Market_US)
 	if err != nil {
 		log.Fatalf("GetPlateSet failed: %v", err)
 	}
