+++ examples/39_option_expiration/main.go
@@ -22,7 +22,7 @@ func main() {
 		log.Fatalf("Connect failed: %v", err)
 	}
 
-	exps, err := client.GetOptionExpirationDate(context.Background(), cli, int32(constant.Market_US), "NVDA")
+	exps, err := client.GetOptionExpirationDate(context.Background(), cli, constant.Market_US, "NVDA")
 	if err != nil {
 		log.Fatalf("GetOptionExpirationDate failed: %v", err)
 	}
